+++ lib/presentation/check_point/check_point_page.dart
@@ -1,5 +1,6 @@
 import 'package:core/core.dart';
 import 'package:dependency/dependency.dart';
+import 'package:design_system/design_system.dart';
 import 'package:flutter_demo_app/data/data.dart';
 import 'package:flutter_demo_app/domain/domain.dart';
 import 'package:flutter_demo_app/presentation/check_point/check_point_controller.dart';